# Docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next.js API (Feature Requests, Profiles, Personas, Contentstack)",
    "description": "API for authentication (Supabase), feature requests (with comments/votes and file uploads), user profiles & personas, Contentstack publishing helpers, and internal log ingestion.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/demo-instructions": {
      "post": {
        "tags": [
          "Contentstack"
        ],
        "summary": "Create & publish a Demo Instruction entry",
        "description": "Creates a `demo_instruction` entry in Contentstack, publishes it, updates a library entry, and republishes the library entry. Requires `PUBLISH_DEMO_INSTRUCTIONS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInstructionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instruction created and published.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoInstructionCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing title or HTML.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Contentstack operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiDemoInstructions"
      }
    },
    "/api/demo-instructions/update": {
      "post": {
        "tags": [
          "Contentstack"
        ],
        "summary": "Update & publish a Demo Instruction entry",
        "description": "Updates an existing `demo_instruction` entry in Contentstack and republishes it. Requires `PUBLISH_DEMO_INSTRUCTIONS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInstructionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instruction updated and published.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoInstructionUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing UID, title, or HTML (or invalid input).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Contentstack operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiDemoInstructionsUpdate"
      }
    },
...
    "/api/feature-requests/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "Feature request ID"
        }
      ],
      "put": {
        "tags": [
          "Feature Requests"
        ],
        "summary": "Update a feature request",
        "description": "Updates title/content/status. If the user lacks `MANAGE_ALL_FEATURE_REQUESTS`, they may only update their own requests. Requires `PUBLISH_FEATURE_REQUESTS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureRequestUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated feature request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureRequest"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing title / invalid input).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not owner or missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "putApiFeatureRequestsId"
      },
      "delete": {
        "tags": [
          "Feature Requests"
        ],
        "summary": "Delete a feature request",
        "description": "Deletes a feature request. If the user lacks `MANAGE_ALL_FEATURE_REQUESTS`, they may only delete their own requests. Requires `PUBLISH_FEATURE_REQUESTS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not owner or missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteApiFeatureRequestsId"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-<project-ref>-auth-token",
        "description": "Supabase SSR session cookie (project-specific name like `sb-<project-ref>-auth-token`)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token (e.g., Supabase access_token)."
      },
      "internalLogSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-log-secret",
        "description": "Shared secret used to authorize internal log ingestion."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "DemoInstructionCreateRequest": {
        "type": "object",
        "required": [
          "title",
          "html"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "html": {
            "type": "string",
            "description": "HTML content for the instruction."
          }
        }
      },
      "DemoInstructionCreateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "instruction_uid": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message",
          "instruction_uid",
          "url"
        ],
        "additionalProperties": false
      },
      "DemoInstructionUpdateRequest": {
        "type": "object",
        "required": [
          "uid",
          "title",
          "html"
        ],
        "properties": {
          "uid": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
      "DemoInstructionUpdateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "entry": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "message",
          "entry"
        ],
        "additionalProperties": false
      },
      "FeatureRequestStatus": {
        "type": "string",
        "enum": [
          "open",
          "planned",
          "in_progress",
          "completed",
          "rejected",
          "closed"
        ],
        "description": "Feature request lifecycle status."
      },
...
      "FeatureRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FeatureRequestStatus"
          },
          "file_url": {
            "type": "string",
            "nullable": true,
            "description": "Path within the `feature-uploads` Supabase storage bucket."
          },
          "signed_file_url": {
            "type": "string",
            "nullable": true,
            "description": "Temporary signed URL for `file_url`."
          },
          "number_of_votes": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "profiles": {
            "type": "object",
            "nullable": true,
            "properties": {
              "username": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "user": {
            "type": "object",
            "nullable": true,
            "properties": {
              "username": {
                "type": "string",
                "nullable": true
              }
            }
          }
        },
        "additionalProperties": true
      },
...
      "FeatureRequestUpdateRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/FeatureRequestStatus"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}